Makes the multitask tests work
